[ refactor ] InternalUserService 추상화 하여 계층간 분리 ( #36 )
