+++ src/main/java/com/project/chaechaeserver/presentation/controller/user/InternalUserController.java
@@ -2,8 +2,8 @@ package com.project.chaechaeserver.presentation.controller.user;
 
 import com.project.chaechaeserver.application.global.dto.ResDTO;
 import com.project.chaechaeserver.application.service.user.InternalUserService;
+import com.project.chaechaeserver.application.service.user.InternalUserServiceImpl;
 import com.project.chaechaeserver.application.response.user.ResInternalUserPostSignupDTO;
-import com.project.chaechaeserver.domain.model.user.constraint.RoleType;
 import com.project.chaechaeserver.presentation.request.user.ReqInternalUserPostSignupDTO;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
